Refactor main.dart to streamline error handling and state management with Riverpod. Introduce installErrorListener for centralized error dialog management. Add WeatherScreen and related widgets for displaying weather conditions and temperatures, enhancing UI structure and responsiveness.
