+++ lib/main.dart
@@ -1,29 +1,15 @@
-import 'dart:async';
-
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:flutter_training/ui/actions/weather_actions.dart';
-import 'package:flutter_training/ui/extension/yumemi_weather_error_extension.dart';
-import 'package:flutter_training/ui/green_widget.dart';
-import 'package:flutter_training/ui/providers/weather_providers.dart';
-import 'package:yumemi_weather/yumemi_weather.dart';
+import 'package:flutter_training/core/bootstrap/install_error_listener.dart';
+import 'package:flutter_training/core/di/overrides.dart';
+import 'package:flutter_training/ui/widget/green_widget.dart';
 
 final navigatorKey = GlobalKey<NavigatorState>();
 
 void main() {
   runApp(
     ProviderScope(
-      overrides: [
-        weatherActionsProvider.overrideWith((ref) {
-          final service = ref.watch(weatherServiceProvider);
-          return WeatherActions(
-            updateWeather: ({required area, required date}) async {
-              service.updateWeather(area: area, date: date);
-            },
-            clearError: service.clearError,
-          );
-        }),
-      ],
+      overrides: overrides,
       child: const MainApp(),
     ),
   );
@@ -34,52 +20,10 @@ class MainApp extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    _listenError(ref);
+    installErrorListener(ref, navigatorKey);
     return MaterialApp(
       navigatorKey: navigatorKey,
       home: const GreenWidget(),
     );
-  }
-
-  void _listenError(WidgetRef ref) {
-    ref.listen(weatherStateProvider.select((state) => state.error), (
-      prev,
-      next,
-    ) {
-      final currentContext = navigatorKey.currentContext;
-      if (currentContext == null || next == null || identical(prev, next)) {
-        return;
-      }
-      WidgetsBinding.instance.addPostFrameCallback((_) {
-        unawaited(_showErrorDialog(currentContext, next, ref));
-      });
-    });
-  }
-
-  Future<void> _showErrorDialog(
-    BuildContext context,
-    YumemiWeatherError error,
-    WidgetRef ref,
-  ) async {
-    if (context.mounted) {
-      await showDialog<void>(
-        context: context,
-        builder: (dialogContext) {
-          return AlertDialog(
-            title: Text(error.title),
-            content: Text(error.description),
-            actions: [
-              TextButton(
-                onPressed: () {
-                  Navigator.of(dialogContext).pop();
-                  ref.read(weatherActionsProvider).clearError();
-                },
-                child: const Text('OK'),
-              ),
-            ],
-          );
-        },
-      );
-    }
   }
 }
